Rename - exercise 재생 페이지 로직, UI 분리하여 가독성 개선
