+++ frontend/src/hooks/exercise/index.ts
@@ -1,0 +1,3 @@
+export { useExercisePlayback } from "./useExercisePlayback";
+export { useSessionLogging } from "./useSessionLogging";
+export type { ExerciseDetail } from "./useExercisePlayback";
